feat: add restaurante rest and handler

+++ src/main/java/com/eatshub/catalog/infrastructure/entrypoints/dto/response/RestaurantResponse.java
@@ -13,6 +13,6 @@ public record RestaurantResponse(
          String logoUrl,
          String closeAt,
          ContactInfo contactInfo,
-         Integer globalRating
+         Double globalRating
 ) {
 }
